+++ datadog_checks_base/datadog_checks/checks/prometheus/base_check.py
@@ -2,7 +2,7 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from .mixins import PrometheusScraperMixin
-from ..base import AgentCheck
+from .. import AgentCheck
 from ...errors import CheckException
 
 from six import string_types
@@ -84,9 +84,7 @@ class GenericPrometheusCheck(AgentCheck):
         - bar
         - foo
     """
-    # This limit is set artificially high for v6.5, it will be reduced to 2000 in 6.6
-    # If you need more than 2000, determine your needed limit by then.
-    DEFAULT_METRIC_LIMIT = 250000
+    DEFAULT_METRIC_LIMIT = 2000
 
     def __init__(self, name, init_config, agentConfig, instances=None, default_instances=None, default_namespace=""):
         super(GenericPrometheusCheck, self).__init__(name, init_config, agentConfig, instances)